Extract creation of layers
