+++ OTAnalytics/plugin_ui/main_application.py
@@ -1,5 +1,7 @@
 import logging
-from typing import Callable, Optional, Sequence
+from typing import Sequence
+
+from plugin_ui.visualization.visualization import VisualizationBuilder
 
 from OTAnalytics.application.analysis.intersect import (
     RunIntersect,
@@ -25,16 +27,10 @@ from OTAnalytics.application.datastore import (
 )
 from OTAnalytics.application.eventlist import SceneActionDetector
 from OTAnalytics.application.logger import logger, setup_logger
-from OTAnalytics.application.plotting import (
-    CachedPlotter,
-    LayeredPlotter,
-    PlottingLayer,
-    TrackBackgroundPlotter,
-)
+from OTAnalytics.application.plotting import LayeredPlotter, PlottingLayer
 from OTAnalytics.application.state import (
     ActionState,
     FlowState,
-    Plotter,
     SectionState,
     SelectedVideoUpdate,
     TrackImageUpdater,
@@ -64,14 +60,6 @@ from OTAnalytics.application.use_cases.generate_flows import (
     GenerateFlows,
     RepositoryFlowIdGenerator,
 )
-from OTAnalytics.application.use_cases.highlight_intersections import (
-    TracksAssignedToGivenFlows,
-    TracksAssignedToSelectedFlows,
-    TracksIntersectingGivenSections,
-    TracksIntersectingSelectedSections,
-    TracksNotIntersectingSelection,
-    TracksOverlapOccurrenceWindow,
-)
 from OTAnalytics.application.use_cases.load_otflow import LoadOtflow
 from OTAnalytics.application.use_cases.load_track_files import LoadTrackFiles
 from OTAnalytics.application.use_cases.reset_project_config import ResetProjectConfig
@@ -99,19 +87,17 @@ from OTAnalytics.application.use_cases.update_project import ProjectUpdater
 from OTAnalytics.application.use_cases.video_repository import ClearAllVideos
 from OTAnalytics.domain.event import EventRepository, SceneEventBuilder
 from OTAnalytics.domain.filter import FilterElementSettingRestorer
-from OTAnalytics.domain.flow import FlowId, FlowRepository
+from OTAnalytics.domain.flow import FlowRepository
 from OTAnalytics.domain.intersect import IntersectImplementation
 from OTAnalytics.domain.progress import ProgressbarBuilder
-from OTAnalytics.domain.section import SectionId, SectionRepository
+from OTAnalytics.domain.section import SectionRepository
 from OTAnalytics.domain.track import (
     ByMaxConfidence,
     PythonTrackDataset,
     TrackFileRepository,
-    TrackIdProvider,
     TrackRepository,
 )
 from OTAnalytics.domain.video import VideoRepository
-from OTAnalytics.plugin_filter.dataframe_filter import DataFrameFilterBuilder
 from OTAnalytics.plugin_intersect.shapely.intersect import ShapelyIntersector
 from OTAnalytics.plugin_intersect.shapely.mapping import ShapelyMapper
 from OTAnalytics.plugin_intersect.simple.cut_tracks_with_sections import (
@@ -148,22 +134,7 @@ from OTAnalytics.plugin_prototypes.eventlist_exporter.eventlist_exporter import 
 )
 from OTAnalytics.plugin_prototypes.track_visualization.track_viz import (
     DEFAULT_COLOR_PALETTE,
-    CachedPandasTrackProvider,
     ColorPaletteProvider,
-    FilterByClassification,
-    FilterById,
-    FilterByOccurrence,
-    FlowLayerPlotter,
-    FlowListObserverWrapper,
-    MatplotlibTrackPlotter,
-    PandasDataFrameProvider,
-    PandasTrackProvider,
-    PandasTracksOffsetProvider,
-    PlotterPrototype,
-    SectionLayerPlotter,
-    SectionListObserverWrapper,
-    TrackGeometryPlotter,
-    TrackStartEndPointPlotter,
 )
 from OTAnalytics.plugin_ui.cli import (
     CliArgumentParser,
@@ -239,19 +210,13 @@ class ApplicationStarter:
         section_state = self._create_section_state(section_repository)
         flow_state = self._create_flow_state()
         road_user_assigner = FilterBySectionEnterEvent(SimpleRoadUserAssigner())
-        cached_pandas_track_provider = self._create_pandas_track_provider(
-            datastore, track_view_state, pulling_progressbar_builder
-        )
-        pandas_data_provider = self._wrap_pandas_track_offset_provider(
-            cached_pandas_track_provider, track_view_state
-        )
         color_palette_provider = ColorPaletteProvider(DEFAULT_COLOR_PALETTE)
         layers = self._create_layers(
             datastore,
             track_view_state,
             flow_state,
             section_state,
-            pandas_data_provider,
+            pulling_progressbar_builder,
             road_user_assigner,
             color_palette_provider,
         )
@@ -598,536 +563,25 @@ class ApplicationStarter:
     def _create_track_view_state(self) -> TrackViewState:
         return TrackViewState()
 
-    def _wrap_plotter_with_cache(
-        self,
-        other: Plotter,
-        datastore: Datastore,
-        tracks: bool,
-        sections: bool,
-        flows: bool,
-    ) -> Plotter:
-        cached_plotter: CachedPlotter = CachedPlotter(other, subjects=[])
-        invalidate = cached_plotter.invalidate_cache
-
-        if tracks:
-            track_repo = datastore._track_repository
-            track_repo.observers.register(invalidate)
-
-        if sections:
-            section_repo = datastore._section_repository
-            section_repo.register_sections_observer(
-                SectionListObserverWrapper(invalidate)
-            )
-            section_repo._section_content_observers.register(invalidate)
-
-        if flows:
-            flow_repo = datastore._flow_repository
-            flow_repo.register_flows_observer(FlowListObserverWrapper(invalidate))
-            flow_repo._flow_content_observers.register(invalidate)
-
-        return cached_plotter
-
-    def _create_pandas_track_provider(
-        self,
-        datastore: Datastore,
-        track_view_state: TrackViewState,
-        progressbar: ProgressbarBuilder,
-    ) -> PandasTrackProvider:
-        dataframe_filter_builder = self._create_dataframe_filter_builder()
-        # return PandasTrackProvider(
-        #     datastore, track_view_state, dataframe_filter_builder, progressbar
-        # )
-        return CachedPandasTrackProvider(
-            datastore, track_view_state, dataframe_filter_builder, progressbar
-        )
-
-    @staticmethod
-    def _wrap_pandas_track_offset_provider(
-        other: PandasDataFrameProvider,
-        track_view_state: TrackViewState,
-    ) -> PandasDataFrameProvider:
-        return PandasTracksOffsetProvider(other, track_view_state)
-
-    def _create_track_geometry_plotter(
-        self,
-        state: TrackViewState,
-        pandas_data_provider: PandasDataFrameProvider,
-        color_palette_provider: ColorPaletteProvider,
-        alpha: float,
-        enable_legend: bool,
-    ) -> Plotter:
-        track_plotter = MatplotlibTrackPlotter(
-            TrackGeometryPlotter(
-                pandas_data_provider,
-                color_palette_provider,
-                alpha=alpha,
-                enable_legend=enable_legend,
-            ),
-        )
-        return PlotterPrototype(state, track_plotter)
-
-    def _create_track_start_end_point_data_provider(
-        self,
-        state: TrackViewState,
-        pandas_data_provider: PandasDataFrameProvider,
-        track_repository: TrackRepository,
-        id_filter: Optional[TrackIdProvider] = None,
-    ) -> FilterById:
-        data_provider = pandas_data_provider
-        data_provider = FilterById(
-            pandas_data_provider,
-            id_filter=TracksOverlapOccurrenceWindow(
-                other=id_filter,
-                track_repository=track_repository,
-                track_view_state=state,
-            ),
-        )
-        return data_provider
-
-    def _create_track_start_end_point_plotter(
-        self,
-        state: TrackViewState,
-        data_provider: PandasDataFrameProvider,
-        color_palette_provider: ColorPaletteProvider,
-        enable_legend: bool,
-    ) -> Plotter:
-        track_plotter = MatplotlibTrackPlotter(
-            TrackStartEndPointPlotter(
-                data_provider,
-                color_palette_provider,
-                enable_legend=enable_legend,
-            ),
-        )
-        return PlotterPrototype(state, track_plotter)
-
-    def _create_tracks_intersecting_selected_sections(
-        self,
-        section_state: SectionState,
-        tracks_intersecting_sections: TracksIntersectingSections,
-        get_sections_by_id: GetSectionsById,
-    ) -> TracksIntersectingSelectedSections:
-        return TracksIntersectingSelectedSections(
-            section_state, tracks_intersecting_sections, get_sections_by_id
-        )
-
-    def _create_tracks_intersecting_sections_filter(
-        self,
-        pandas_data_provider: PandasDataFrameProvider,
-        tracks_intersecting_sections: TracksIntersectingSections,
-        get_sections_by_id: GetSectionsById,
-    ) -> Callable[[SectionId], PandasDataFrameProvider]:
-        return lambda section: FilterById(
-            pandas_data_provider,
-            TracksIntersectingGivenSections(
-                [section], tracks_intersecting_sections, get_sections_by_id
-            ),
-        )
-
-    def _create_highlight_tracks_intersecting_section_factory(
-        self,
-        state: TrackViewState,
-        pandas_data_provider_factory: Callable[[SectionId], PandasDataFrameProvider],
-        color_palette_provider: ColorPaletteProvider,
-        alpha: float,
-        enable_legend: bool,
-    ) -> Callable[[SectionId], Plotter]:
-        return lambda section: self._create_track_geometry_plotter(
-            state,
-            pandas_data_provider_factory(section),
-            color_palette_provider,
-            alpha,
-            enable_legend,
-        )
-
-    def _create_cached_section_layer_plotter(
-        self,
-        plotter_factory: Callable[[SectionId], Plotter],
-        section_state: SectionState,
-        section_repository: SectionRepository,
-        track_repository: TrackRepository,
-    ) -> Plotter:
-        return SectionLayerPlotter(
-            plotter_factory, section_state, section_repository, track_repository
-        )
-
-    def _create_track_highlight_geometry_plotter_not_intersecting(
-        self,
-        state: TrackViewState,
-        tracks_not_intersecting_sections: TracksNotIntersectingSelection,
-        pandas_track_provider: PandasDataFrameProvider,
-        color_palette_provider: ColorPaletteProvider,
-        enable_legend: bool,
-    ) -> Plotter:
-        filter_by_id = FilterById(
-            pandas_track_provider, id_filter=tracks_not_intersecting_sections
-        )
-        return self._create_track_geometry_plotter(
-            state,
-            filter_by_id,
-            color_palette_provider,
-            alpha=1,
-            enable_legend=enable_legend,
-        )
-
-    def _create_tracks_start_end_point_intersecting_given_sections_filter(
-        self,
-        state: TrackViewState,
-        track_repository: TrackRepository,
-        pandas_data_provider: PandasDataFrameProvider,
-        tracks_intersecting_sections: TracksIntersectingSections,
-        get_sections_by_id: GetSectionsById,
-    ) -> Callable[[SectionId], PandasDataFrameProvider]:
-        return lambda section: FilterById(
-            pandas_data_provider,
-            id_filter=TracksOverlapOccurrenceWindow(
-                other=TracksIntersectingGivenSections(
-                    [section],
-                    tracks_intersecting_sections,
-                    get_sections_by_id,
-                ),
-                track_repository=track_repository,
-                track_view_state=state,
-            ),
-        )
-
-    def _create_start_end_point_intersecting_section_factory(
-        self,
-        state: TrackViewState,
-        pandas_data_provider_factory: Callable[[SectionId], PandasDataFrameProvider],
-        color_palette_provider: ColorPaletteProvider,
-        enable_legend: bool,
-    ) -> Callable[[SectionId], Plotter]:
-        return lambda section: self._create_track_start_end_point_plotter(
-            state,
-            pandas_data_provider_factory(section),
-            color_palette_provider,
-            enable_legend,
-        )
-
-    def _create_start_end_point_tracks_not_intersecting_sections_plotter(
-        self,
-        state: TrackViewState,
-        tracks_not_intersecting_sections: TracksNotIntersectingSelection,
-        pandas_track_provider: PandasDataFrameProvider,
-        track_repository: TrackRepository,
-        color_palette_provider: ColorPaletteProvider,
-        enable_legend: bool,
-    ) -> Plotter:
-        return self._create_track_start_end_point_plotter(
-            state,
-            self._create_track_start_end_point_data_provider(
-                state,
-                pandas_track_provider,
-                track_repository,
-                tracks_not_intersecting_sections,
-            ),
-            color_palette_provider,
-            enable_legend,
-        )
-
-    def _create_tracks_assigned_to_flows_filter(
-        self,
-        pandas_data_provider: PandasDataFrameProvider,
-        assigner: RoadUserAssigner,
-        event_repository: EventRepository,
-        flow_repository: FlowRepository,
-    ) -> Callable[[FlowId], PandasDataFrameProvider]:
-        return lambda flow: FilterById(
-            pandas_data_provider,
-            TracksAssignedToGivenFlows(
-                assigner,
-                event_repository,
-                flow_repository,
-                [flow],
-            ),
-        )
-
-    def _create_highlight_tracks_assigned_to_flows_factory(
-        self,
-        state: TrackViewState,
-        pandas_data_provider_factory: Callable[[FlowId], PandasDataFrameProvider],
-        color_palette_provider: ColorPaletteProvider,
-        alpha: float,
-        enable_legend: bool,
-    ) -> Callable[[FlowId], Plotter]:
-        return lambda flow: self._create_track_geometry_plotter(
-            state,
-            pandas_data_provider_factory(flow),
-            color_palette_provider,
-            alpha,
-            enable_legend,
-        )
-
-    def _create_highlight_tracks_assigned_to_flow(
-        self,
-        plotter_factory: Callable[[FlowId], Plotter],
-        flow_state: FlowState,
-        flow_repository: FlowRepository,
-        track_repository: TrackRepository,
-    ) -> Plotter:
-        return FlowLayerPlotter(
-            plotter_factory, flow_state, flow_repository, track_repository
-        )
-
-    def _create_highlight_tracks_not_assigned_to_flow(
-        self,
-        state: TrackViewState,
-        pandas_track_provider: PandasDataFrameProvider,
-        color_palette_provider: ColorPaletteProvider,
-        tracks_assigned_to_flow: TracksAssignedToSelectedFlows,
-        track_repository: TrackRepository,
-        enable_legend: bool,
-    ) -> Plotter:
-        tracks_not_assigned_to_flow = TracksNotIntersectingSelection(
-            tracks_assigned_to_flow, track_repository
-        )
-        filter_by_id = FilterById(
-            pandas_track_provider, id_filter=tracks_not_assigned_to_flow
-        )
-        return self._create_track_geometry_plotter(
-            state,
-            filter_by_id,
-            color_palette_provider,
-            alpha=1,
-            enable_legend=enable_legend,
-        )
-
     def _create_layers(
         self,
         datastore: Datastore,
         track_view_state: TrackViewState,
         flow_state: FlowState,
         section_state: SectionState,
-        pandas_data_provider: PandasDataFrameProvider,
+        pulling_progressbar_builder: ProgressbarBuilder,
         road_user_assigner: RoadUserAssigner,
         color_palette_provider: ColorPaletteProvider,
     ) -> Sequence[PlottingLayer]:
-        background_image_plotter = TrackBackgroundPlotter(track_view_state, datastore)
-        data_provider_all_filters = FilterByClassification(
-            FilterByOccurrence(
-                pandas_data_provider,
-                track_view_state,
-                self._create_dataframe_filter_builder(),
-            ),
+        return VisualizationBuilder(datastore).build(
+            datastore,
             track_view_state,
-            self._create_dataframe_filter_builder(),
-        )
-        data_provider_class_filter = FilterByClassification(
-            pandas_data_provider,
-            track_view_state,
-            self._create_dataframe_filter_builder(),
-        )
-        track_geometry_plotter = self._create_track_geometry_plotter(
-            track_view_state,
-            data_provider_all_filters,
-            color_palette_provider,
-            alpha=0.5,
-            enable_legend=True,
-        )
-        tracks_intersecting_sections = self._create_tracks_intersecting_sections(
-            GetTracksWithoutSingleDetections(datastore._track_repository),
-            ShapelyIntersector(),
-        )
-        tracks_intersecting_selected_sections = (
-            self._create_tracks_intersecting_selected_sections(
-                section_state,
-                tracks_intersecting_sections,
-                GetSectionsById(datastore._section_repository),
-            )
-        )
-        tracks_not_intersecting_sections = TracksNotIntersectingSelection(
-            tracks_intersecting_selected_sections, datastore._track_repository
-        )
-
-        get_sections_by_id = GetSectionsById(datastore._section_repository)
-        tracks_intersecting_sections_filter = (
-            self._create_tracks_intersecting_sections_filter(
-                data_provider_all_filters,
-                tracks_intersecting_sections,
-                get_sections_by_id,
-            )
-        )
-
-        highlight_tracks_intersecting_sections = (
-            self._create_cached_section_layer_plotter(
-                self._create_highlight_tracks_intersecting_section_factory(
-                    track_view_state,
-                    tracks_intersecting_sections_filter,
-                    color_palette_provider,
-                    alpha=1,
-                    enable_legend=False,
-                ),
-                section_state,
-                datastore._section_repository,
-                datastore._track_repository,
-            )
-        )
-        highlight_tracks_not_intersecting_sections = (
-            self._create_track_highlight_geometry_plotter_not_intersecting(
-                track_view_state,
-                tracks_not_intersecting_sections,
-                data_provider_all_filters,
-                color_palette_provider,
-                enable_legend=False,
-            )
-        )
-        start_end_points_intersecting = self._create_cached_section_layer_plotter(
-            self._create_start_end_point_intersecting_section_factory(
-                track_view_state,
-                self._create_tracks_start_end_point_intersecting_given_sections_filter(
-                    track_view_state,
-                    datastore._track_repository,
-                    data_provider_class_filter,
-                    tracks_intersecting_sections,
-                    get_sections_by_id,
-                ),
-                color_palette_provider,
-                enable_legend=False,
-            ),
-            section_state,
-            datastore._section_repository,
-            datastore._track_repository,
-        )
-
-        start_end_points_tracks_not_intersecting_sections = (
-            self._create_start_end_point_tracks_not_intersecting_sections_plotter(
-                track_view_state,
-                tracks_not_intersecting_sections,
-                data_provider_class_filter,
-                datastore._track_repository,
-                color_palette_provider,
-                enable_legend=False,
-            )
-        )
-        track_start_end_point_plotter = self._create_track_start_end_point_plotter(
-            track_view_state,
-            self._create_track_start_end_point_data_provider(
-                track_view_state,
-                data_provider_class_filter,
-                datastore._track_repository,
-            ),
-            color_palette_provider,
-            enable_legend=False,
-        )
-        tracks_assigned_to_flow = TracksAssignedToSelectedFlows(
-            road_user_assigner,
-            datastore._event_repository,
-            datastore._flow_repository,
             flow_state,
+            section_state,
+            pulling_progressbar_builder,
+            road_user_assigner,
+            color_palette_provider,
         )
-
-        highlight_tracks_assigned_to_flow = (
-            self._create_highlight_tracks_assigned_to_flow(
-                self._create_highlight_tracks_assigned_to_flows_factory(
-                    track_view_state,
-                    self._create_tracks_assigned_to_flows_filter(
-                        data_provider_all_filters,
-                        road_user_assigner,
-                        datastore._event_repository,
-                        datastore._flow_repository,
-                    ),
-                    color_palette_provider,
-                    alpha=1,
-                    enable_legend=False,
-                ),
-                flow_state,
-                datastore._flow_repository,
-                datastore._track_repository,
-            )
-        )
-        highlight_tracks_not_assigned_to_flow = (
-            self._create_highlight_tracks_not_assigned_to_flow(
-                track_view_state,
-                data_provider_all_filters,
-                color_palette_provider,
-                tracks_assigned_to_flow,
-                datastore._track_repository,
-                enable_legend=False,
-            )
-        )
-        background = PlottingLayer("Background", background_image_plotter, enabled=True)
-        all_tracks_layer = PlottingLayer(
-            "Show all tracks",
-            self._wrap_plotter_with_cache(
-                track_geometry_plotter,
-                datastore,
-                tracks=True,
-                sections=False,
-                flows=False,
-            ),
-            enabled=False,
-        )
-        highlight_tracks_intersecting_sections_layer = PlottingLayer(
-            "Highlight tracks intersecting sections",
-            highlight_tracks_intersecting_sections,
-            enabled=False,
-        )
-        highlight_tracks_not_intersecting_sections_layer = PlottingLayer(
-            "Highlight tracks not intersecting sections",
-            self._wrap_plotter_with_cache(
-                highlight_tracks_not_intersecting_sections,
-                datastore,
-                tracks=True,
-                sections=True,
-                flows=False,
-            ),
-            enabled=False,
-        )
-        start_end_points_tracks_intersecting_sections_layer = PlottingLayer(
-            "Show start and end point of tracks intersecting sections",
-            start_end_points_intersecting,
-            enabled=False,
-        )
-        start_end_points_tracks_not_intersecting_sections_layer = PlottingLayer(
-            "Show start and end point of tracks not intersecting sections",
-            self._wrap_plotter_with_cache(
-                start_end_points_tracks_not_intersecting_sections,
-                datastore,
-                tracks=True,
-                sections=True,
-                flows=False,
-            ),
-            enabled=False,
-        )
-        start_end_point_layer = PlottingLayer(
-            "Show start and end point",
-            self._wrap_plotter_with_cache(
-                track_start_end_point_plotter,
-                datastore,
-                tracks=True,
-                sections=False,
-                flows=False,
-            ),
-            enabled=False,
-        )
-        highlight_tracks_assigned_to_flow_layer = PlottingLayer(
-            "Highlight tracks assigned to flow",
-            highlight_tracks_assigned_to_flow,
-            enabled=False,
-        )
-        highlight_tracks_not_assigned_to_flow_layer = PlottingLayer(
-            "Highlight tracks not assigned to flow",
-            self._wrap_plotter_with_cache(
-                highlight_tracks_not_assigned_to_flow,
-                datastore,
-                tracks=True,
-                sections=True,
-                flows=True,
-            ),
-            enabled=False,
-        )
-
-        return [
-            background,
-            all_tracks_layer,
-            highlight_tracks_intersecting_sections_layer,
-            highlight_tracks_not_intersecting_sections_layer,
-            start_end_point_layer,
-            start_end_points_tracks_intersecting_sections_layer,
-            start_end_points_tracks_not_intersecting_sections_layer,
-            highlight_tracks_assigned_to_flow_layer,
-            highlight_tracks_not_assigned_to_flow_layer,
-        ]
 
     @staticmethod
     def _create_section_state(section_repository: SectionRepository) -> SectionState:
@@ -1186,9 +640,6 @@ class ApplicationStarter:
 
     def _create_action_state(self) -> ActionState:
         return ActionState()
-
-    def _create_dataframe_filter_builder(self) -> DataFrameFilterBuilder:
-        return DataFrameFilterBuilder()
 
     def _create_filter_element_setting_restorer(self) -> FilterElementSettingRestorer:
         return FilterElementSettingRestorer()
